+++ node_package/src/handleError.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { renderToString } from './reactApis.cts';
+import { renderToString } from './ReactDOMServer.cts';
 import type { ErrorOptions } from './types/index.ts';
 
 function handleRenderFunctionIssue(options: ErrorOptions): string {
